Export types via 'PixelEventTypes' entrypoint
